+++ app/src/components/common/Button/index.ts
@@ -1,2 +1,3 @@
 export * from './Button';
+export * from './Clickable';
 export * from './StartButton';
