+++ app/src/components/common/index.ts
@@ -1,6 +1,5 @@
 export * from './Avatar';
 export * from './Button';
-export * from './Clickable';
 export * from './DeferredComponent';
 export * from './Divider';
 export * from './Image';
